fixbug datetime and add option without_remote_database

+++ registrator_romania/backend/strategies_registration.py
@@ -50,6 +50,7 @@ class StrategyWithoutProxy:
         use_shuffle: bool = True,
         logging: bool = True,
         residental_proxy_url: str = None,
+        without_remote_database: bool = False
     ) -> None:
         if not stop_when:
             stop_when = [9, 2]
@@ -64,6 +65,7 @@ class StrategyWithoutProxy:
         self._use_shuffle = use_shuffle
         self._logging = logging
         self._residental_proxy_url = residental_proxy_url
+        self._without_remote_database = without_remote_database
 
     async def start(self):
         if self._users_data:
@@ -137,9 +139,10 @@ class StrategyWithoutProxy:
 
         if api.is_success_registration(html):
             try:
-                async with asyncio.timeout(5):
-                    async with self._db as db:
-                        await db.remove_user(user_data)
+                if not self._without_remote_database:
+                    async with asyncio.timeout(5):
+                        async with self._db as db:
+                            await db.remove_user(user_data)
             except asyncio.TimeoutError:
                 pass
             except Exception as e:
@@ -159,9 +162,10 @@ class StrategyWithoutProxy:
             if error.count("Deja a fost înregistrată o programare"):
                 await queue.put((user_data.copy(), html))
                 try:
-                    async with asyncio.timeout(5):
-                        async with self._db as db:
-                            await db.remove_user(user_data)
+                    if not self._without_remote_database:
+                        async with asyncio.timeout(5):
+                            async with self._db as db:
+                                await db.remove_user(user_data)
                 except asyncio.TimeoutError:
                     pass
                 except Exception as e:
@@ -198,6 +202,7 @@ class StrategyWithoutProxy:
         reg_dt = self._registration_date
         successfully_registered = []
         queue = asyncio.Queue()
+        users_for_registrate = []
 
         now = self._get_dt_now()
         dirname = f"registrations_{reg_dt.strftime("%d.%m.%Y")}"
@@ -285,10 +290,11 @@ class StrategyWithoutProxy:
     async def update_users_list(self):
         while True:
             try:
-                async with self._db as db:
-                    self._users_data = await db.get_users_by_reg_date(
-                        self._registration_date
-                    )
+                if not self._without_remote_database:
+                    async with self._db as db:
+                        self._users_data = await db.get_users_by_reg_date(
+                            self._registration_date
+                        )
             except asyncio.TimeoutError:
                 pass
             except Exception as e:
@@ -335,11 +341,12 @@ class StrategyWithoutProxy:
 
     async def add_users_to_db(self):
         try:
-            async with self._db as db:
-                for user_data in self._users_data:
-                    await db.add_user(
-                        user_data, registration_date=self._registration_date
-                    )
+            if not self._without_remote_database:
+                async with self._db as db:
+                    for user_data in self._users_data:
+                        await db.add_user(
+                            user_data, registration_date=self._registration_date
+                        )
         except asyncio.TimeoutError:
             pass
         except Exception as e:
